metrics-collector/requirements.txt
psycopg2-binary==2.9.9
docker==7.0.0
psutil==5.9.8
pandas==2.1.4
requests==2.31.0
schedule==1.2.1